Prettied up panel UIs, added solar panel UI, made it possible to jump using portal UI, moved time icon to the right place, more UI fixes, fix sizing bug in sprites

+++ src/com/evervoid/client/views/EverUIView.java
@@ -241,7 +241,9 @@ public abstract class EverUIView extends EverView
 
 	public void switchUI(final UIControl newRoot)
 	{
-		deleteUI();
-		pushUI(newRoot);
+		if (aTopUI != newRoot) {
+			deleteUI();
+			pushUI(newRoot);
+		}
 	}
 }
